ECF-RP-313: refactor routes
