+++ app/views/admin/administrators/administrators/index.html.erb
@@ -1,7 +1,7 @@
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
     <h1 class="govuk-heading-xl">Administrators</h1>
-    <%= govuk_link_to "Create a new administrator", new_admin_administrators_administrator_path, button: true %>
+    <%= govuk_link_to "Create a new administrator", new_admin_administrator_path, button: true %>
     <table class="govuk-table">
       <thead>
         <tr class="govuk-table__row">
