+++ app/views/admin/administrators/administrators/new.html.erb
@@ -1,13 +1,13 @@
 <% content_for :before_content, govuk_back_link(
   text: "Back",
-  href: admin_administrators_administrators_path)
+  href: admin_administrators_path)
 %>
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
     <h1 class="govuk-heading-xl">Create a new administrator user</h1>
     <p class="govuk-body">This should be someone from the Department for Education</p>
-    <%= form_for @user, url: admin_administrators_administrators_path, method: :post do |f| %>
+    <%= form_for @user, url: admin_administrators_path, method: :post do |f| %>
       <%= f.govuk_error_summary %>
       <%= f.govuk_text_field(
             :full_name,
