--- app/views/admin/administrators/success/create.html.erb
@@ -1,8 +1,0 @@
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <%= govuk_panel title: "Administrator created", body: nil %>
-
-    <%= govuk_link_to "Add another administrator", new_admin_administrators_administrator_path, button: true %>
-    <%= govuk_link_to "View all administrators", admin_administrators_administrators_path, button: true, class: "govuk-button--secondary" %>
-  </div>
-</div>
