+++ spec/requests/admin/administrators/administrators_spec.rb
@@ -5,6 +5,7 @@ require "rails_helper"
 RSpec.describe "Admin::Administrators::Administrators", type: :request do
   let(:name) { Faker::Name.name }
   let(:email) { Faker::Internet.email }
+  let(:new_user) { User.find_by(email: email) }
 
   before do
     user = create(:user, :admin)
@@ -60,6 +61,34 @@ RSpec.describe "Admin::Administrators::Administrators", type: :request do
     end
   end
 
+  describe "POST /admin/administrators/new/success" do
+    it "creates a new user" do
+      expect { create_new_user }.to change { User.count }.by 1
+    end
+
+    it "creates a new admin profile" do
+      expect { create_new_user }.to change { AdminProfile.count }.by 1
+    end
+
+    it "confirms the new user" do
+      given_a_user_is_created
+
+      expect(new_user.confirmed?).to be true
+    end
+
+    it "makes the new user an admin" do
+      given_a_user_is_created
+
+      expect(new_user.admin?).to be true
+    end
+
+    it "renders a success message" do
+      given_a_user_is_created
+
+      expect(response).to render_template("admin/administrators/administrators/success")
+    end
+  end
+
 private
 
   def given_i_have_previously_submitted_values(name, email)
@@ -68,4 +97,13 @@ private
       email: email,
     } }
   end
+
+  def create_new_user
+    post "/admin/administrators/new/success", params: { user: {
+      full_name: name,
+      email: email,
+    } }
+  end
+
+  alias_method :given_a_user_is_created, :create_new_user
 end
